Refactored NumberUtils adding safe cast methods
